+++ packages/react/src/components/form/authenticating/password.tsx
@@ -131,7 +131,7 @@ function getTextKeys(
  * Renders a form that enables authentication via a password.
  * Handles retries in case of submitting an invalid/incorrect password.
  */
-export const PasswordState = ({ flowState }: Props) => {
+export const PasswordState = ({ flowState, performLogin }: Props) => {
   const { sid } = useSlashID();
   const { text } = useConfiguration();
   const {
@@ -242,13 +242,15 @@ export const PasswordState = ({ flowState }: Props) => {
     sid?.subscribe("incorrectPasswordSubmitted", onIncorrectPassword);
     sid?.subscribe("invalidPasswordSubmitted", onInvalidPassword);
 
+    performLogin();
+
     return () => {
       sid?.unsubscribe("passwordSetReady", onSetPassword);
       sid?.unsubscribe("passwordVerifyReady", onVerifyPassword);
       sid?.unsubscribe("incorrectPasswordSubmitted", onIncorrectPassword);
       sid?.unsubscribe("invalidPasswordSubmitted", onInvalidPassword);
     };
-  }, [setError, sid, text, values]);
+  }, [formState, performLogin, setError, sid, text, values]);
 
   return (
     <>