+++ packages/react/src/context/slash-id-context.tsx
@@ -9,6 +9,7 @@ import {
 } from "react";
 
 import {
+  AnonymousUser,
   PersonHandleType,
   SlashID,
   SlashIDEnvironment,
@@ -23,6 +24,8 @@ import {
 import { LogIn, MFA, Recover } from "../domain/types";
 import { SDKState } from "../domain/sdk-state";
 import { applyMiddleware } from "../middleware";
+import { isAnonymous } from "../domain/user";
+import { sequence } from "../components/utils";
 
 export type StorageOption = "memory" | "localStorage" | "cookie";
 
@@ -52,6 +55,20 @@ export interface SlashIDProviderProps {
    */
   sdkUrl?: string;
   analyticsEnabled?: boolean;
+  /**
+   * Anonymous users allow you to perform operations
+   * on users who have not signed-up or logged-in,
+   * using the same API as regular users.
+   *
+   * Use anonymous users to read & set user attributes, store
+   * GDPR consent, track sign-up conversions, and build a
+   * fingerprint of users who simply never sign-up.
+   *
+   * When the anonymous users signs up, or signs in, their
+   * anonymous user is upgraded to a full user, and their
+   * pre-login data is transferred.
+   */
+  anonymousUsersEnabled?: boolean;
   themeProps?: ThemeProps;
   children: ReactNode;
 }
@@ -59,6 +76,7 @@ export interface SlashIDProviderProps {
 export interface ISlashIDContext {
   sid: SlashID | undefined;
   user: User | undefined;
+  anonymousUser: AnonymousUser | undefined;
   sdkState: SDKState;
   logOut: () => undefined;
   logIn: LogIn;
@@ -71,6 +89,7 @@ export interface ISlashIDContext {
 export const initialContextValue = {
   sid: undefined,
   user: undefined,
+  anonymousUser: undefined,
   sdkState: "initial" as const,
   logOut: () => undefined,
   logIn: () => Promise.reject("NYI"),
@@ -107,6 +126,7 @@ export const SlashIDProvider = ({
   baseApiUrl,
   sdkUrl,
   analyticsEnabled,
+  anonymousUsersEnabled = false,
   themeProps,
   children,
 }: SlashIDProviderProps) => {
@@ -114,6 +134,9 @@ export const SlashIDProvider = ({
   const [token, setToken] = useState(initialToken);
   const [state, setState] = useState<SDKState>(initialContextValue.sdkState);
   const [user, setUser] = useState<User | undefined>(undefined);
+  const [anonymousUser, setAnonymousUser] = useState<AnonymousUser | undefined>(
+    undefined
+  );
   const storageRef = useRef<Storage | undefined>(undefined);
   const sidRef = useRef<SlashID | undefined>(undefined);
 
@@ -143,12 +166,6 @@ export const SlashIDProvider = ({
       setUser(newUser);
       storageRef.current?.setItem(STORAGE_TOKEN_KEY, newUser.token);
 
-      try {
-        sidRef.current?.getAnalytics().identify(newUser);
-      } catch {
-        // fail silently
-      }
-
       if (newUser.oid !== oid) {
         __switchOrganizationInContext({ oid: newUser.oid });
       }
@@ -156,15 +173,45 @@ export const SlashIDProvider = ({
     [state, __switchOrganizationInContext, oid]
   );
 
+  const storeAnonymousUser = useCallback(
+    (anonUser: AnonymousUser) => {
+      if (!anonymousUsersEnabled || state === "initial") {
+        return;
+      }
+
+      setAnonymousUser(anonUser);
+      storageRef.current?.setItem(STORAGE_TOKEN_KEY, anonUser.token);
+
+      try {
+        // core SDK does not track anon users automatically
+        sidRef.current?.getAnalytics().identify(anonUser);
+      } catch {
+        // fail silently
+      }
+    },
+    [anonymousUsersEnabled, state]
+  );
+
+  const clearAnonymousUser = useCallback(() => {
+    if (!anonymousUsersEnabled || state === "initial") {
+      return;
+    }
+
+    storageRef.current?.removeItem(STORAGE_TOKEN_KEY);
+
+    setAnonymousUser(undefined);
+  }, [anonymousUsersEnabled, state]);
+
   const logOut = useCallback((): undefined => {
     if (state === "initial") {
       return;
     }
 
-    storageRef.current?.removeItem(STORAGE_TOKEN_KEY);
-    if (!user) {
+    if (!user || isAnonymous(user)) {
       return;
     }
+
+    storageRef.current?.removeItem(STORAGE_TOKEN_KEY);
 
     try {
       sidRef.current?.getAnalytics().logout();
@@ -201,25 +248,60 @@ export const SlashIDProvider = ({
                 value: handle.value,
               };
 
-        const user = await sid
+        const shouldUpgradeUser = anonymousUsersEnabled && anonymousUser;
+
+        if (shouldUpgradeUser) {
+          const upgradedUser = await anonymousUser
+            // @ts-expect-error TODO make the identifier optional
+            .id(identifier, factor)
+            .then(async (user) => {
+              return applyMiddleware({ user, sid, middleware });
+            });
+
+          storeUser(upgradedUser);
+
+          return upgradedUser;
+        }
+
+        const newUser = await sid
           // @ts-expect-error TODO make the identifier optional
           .id(oid, identifier, factor)
           .then(async (user) => {
             return applyMiddleware({ user, sid, middleware });
           });
 
-        storeUser(user);
-        return user;
+        if (anonymousUsersEnabled && shouldUpgradeUser) {
+          clearAnonymousUser();
+        }
+
+        storeUser(newUser);
+
+        return newUser;
       } catch (e) {
         logOut();
         throw e;
       }
     },
-    [oid, state, storeUser, logOut]
+    [
+      state,
+      anonymousUsersEnabled,
+      anonymousUser,
+      oid,
+      storeUser,
+      clearAnonymousUser,
+      logOut,
+    ]
   );
 
   const mfa = useCallback<MFA>(
     async ({ handle, factor }) => {
+      if (anonymousUsersEnabled && anonymousUser) {
+        console.warn(
+          "Anonymous users cannot perform MFA, please log in first."
+        );
+        return;
+      }
+
       if (state === "initial" || !user) {
         return;
       }
@@ -229,7 +311,7 @@ export const SlashIDProvider = ({
       storeUser(userAfterMfa);
       return userAfterMfa;
     },
-    [state, user, storeUser]
+    [anonymousUsersEnabled, anonymousUser, state, user, storeUser]
   );
 
   /**
@@ -244,16 +326,24 @@ export const SlashIDProvider = ({
     [state]
   );
 
-  const validateToken = useCallback(async (token: string): Promise<boolean> => {
-    const tokenUser = new User(token, sidRef.current!);
-    try {
-      const ret = await tokenUser.validateToken();
-      return ret.valid;
-    } catch (e) {
-      console.error(e);
-      return false;
-    }
-  }, []);
+  const validateToken = useCallback(
+    async (token: string): Promise<boolean> => {
+      const tokenUser = new User(token, sidRef.current!);
+
+      if (tokenUser?.anonymous && !anonymousUsersEnabled) {
+        return false;
+      }
+
+      try {
+        const ret = await tokenUser.validateToken();
+        return ret.valid;
+      } catch (e) {
+        console.error(e);
+        return false;
+      }
+    },
+    [anonymousUsersEnabled]
+  );
 
   useEffect(() => {
     if (state === "initial") {
@@ -281,6 +371,28 @@ export const SlashIDProvider = ({
     environment,
   ]);
 
+  const createAndStoreUserFromToken = useCallback(
+    (token: string): User | AnonymousUser | null => {
+      const sid = sidRef.current!;
+      const newUser = new User(token, sid);
+
+      if (newUser.anonymous && !anonymousUsersEnabled) {
+        return null;
+      }
+
+      if (newUser.anonymous) {
+        const anonUser = new AnonymousUser(token, sid);
+
+        storeAnonymousUser(anonUser);
+        return anonUser;
+      }
+
+      storeUser(newUser);
+      return newUser;
+    },
+    [anonymousUsersEnabled, storeAnonymousUser, storeUser]
+  );
+
   useEffect(() => {
     if (state !== "loaded") {
       return;
@@ -289,61 +401,97 @@ export const SlashIDProvider = ({
     const sid = sidRef.current!;
     const storage = storageRef.current!;
 
-    const loginDirectIdIfPresent = async () => {
+    const loginWithInitialToken = async () => {
+      const isTokenValid = token && (await validateToken(token));
+      if (!isTokenValid) return null;
+
+      return createAndStoreUserFromToken(token);
+    };
+
+    const loginWithDirectID = async () => {
       try {
-        const tempUser = await sid.getUserFromURL();
-        if (tempUser) {
-          storeUser(new User(tempUser.token, sidRef.current!));
-          return true;
-        } else {
-          return false;
-        }
+        const userFromURL = await sid.getUserFromURL();
+        if (!userFromURL) return null;
+
+        const { token: tokenFromURL } = userFromURL;
+
+        return createAndStoreUserFromToken(tokenFromURL);
       } catch (e) {
         console.error(e);
-        return false;
+        return null;
       }
     };
 
-    const loginStoredToken = async (): Promise<boolean> => {
-      const storedToken = storage.getItem(STORAGE_TOKEN_KEY);
-      if (storedToken) {
-        const isValidToken = await validateToken(storedToken);
-        if (!isValidToken) {
-          storage.removeItem(STORAGE_TOKEN_KEY);
-          return false;
-        }
+    const loginWithTokenFromStorage = async () => {
+      const tokenFromStorage = storage.getItem(STORAGE_TOKEN_KEY);
 
-        storeUser(new User(storedToken, sidRef.current!));
-        return true;
-      } else {
-        return false;
+      const isValidToken =
+        tokenFromStorage && (await validateToken(tokenFromStorage));
+
+      if (!isValidToken) {
+        storage.removeItem(STORAGE_TOKEN_KEY);
+
+        return null;
       }
+
+      const user = createAndStoreUserFromToken(tokenFromStorage);
+
+      if (user) {
+        try {
+          // in all other cases the core SDK will handle this
+          // here we just recreate the user object based on the preexisting token
+          // no event is emitted on the SDK side because of that
+          sidRef.current?.getAnalytics().identify(user);
+        } catch {
+          // fail silently
+        }
+      }
+
+      return user;
     };
 
-    const tryImmediateLogin = async () => {
-      if (token) {
-        storeUser(new User(token, sidRef.current!));
-      } else {
-        const isDone = await loginDirectIdIfPresent();
+    const createAnonymousUser = async () => {
+      if (!anonymousUsersEnabled) return null;
 
-        if (!isDone) {
-          await loginStoredToken();
-        }
-      }
+      const anonUser = await sid.createAnonymousUser();
 
-      setState("ready");
+      storeAnonymousUser(anonUser);
+
+      return anonUser;
     };
 
     setState("retrievingToken");
 
-    tryImmediateLogin();
-  }, [state, token, storeUser, validateToken]);
+    sequence(
+      [
+        loginWithInitialToken,
+        loginWithDirectID,
+        loginWithTokenFromStorage,
+        createAnonymousUser,
+      ],
+      {
+        until: (value) => value !== null,
+        then: () => {
+          setState("ready");
+        },
+      }
+    );
+  }, [
+    anonymousUsersEnabled,
+    createAndStoreUserFromToken,
+    state,
+    storeAnonymousUser,
+    storeUser,
+    token,
+    validateToken,
+  ]);
 
   const contextValue = useMemo(() => {
     if (state === "initial") {
       return {
         sid: undefined,
         user,
+        anonymousUser,
         sdkState: state,
         logOut,
         logIn,
@@ -357,6 +505,7 @@ export const SlashIDProvider = ({
     return {
       sid: sidRef.current!,
       user,
+      anonymousUser,
       sdkState: state,
       logOut,
       logIn,
@@ -368,6 +517,7 @@ export const SlashIDProvider = ({
   }, [
     state,
     user,
+    anonymousUser,
     logOut,
     logIn,
     mfa,