+++ packages/react/src/hooks/use-slash-id.test.tsx
@@ -6,11 +6,12 @@ import {
   TEST_PERSON_ID,
   createTestUser,
 } from "../components/test-utils";
+import { BASE_API_URL_CUSTOM } from "../mocks/handlers";
 
 const TestComponent = () => {
   const { user } = useSlashID();
 
-  if (!user) {
+  if (!user || user.anonymous) {
     return <div>Not logged in</div>;
   }
 
@@ -32,8 +33,10 @@ const TestEnvironmentComponent = () => {
 
 describe("useSlashID", () => {
   test("should return a user instance when a valid initial token is passed to the SlashIDProvider", async () => {
+    const { token } = createTestUser();
+
     render(
-      <SlashIDProvider initialToken={createTestUser().token} oid={TEST_ORG_ID}>
+      <SlashIDProvider initialToken={token} oid={TEST_ORG_ID}>
         <TestComponent />
       </SlashIDProvider>
     );
@@ -46,7 +49,7 @@ describe("useSlashID", () => {
 
   test("should use proper custom environment", async () => {
     const customEnv = {
-      baseURL: "https://custom.base.url",
+      baseURL: BASE_API_URL_CUSTOM,
       sdkURL: "https://custom.sdk.url",
     };