+++ apps/react-vite/CHANGELOG.md
@@ -1,5 +1,30 @@
 # react-vite
 
+## 0.0.66
+
+### Patch Changes
+
+- Updated dependencies [25960c7]
+  - @slashid/react@1.24.1
+  - demo-form@0.0.59
+
+## 0.0.65
+
+### Patch Changes
+
+- Updated dependencies [689a2cd]
+- Updated dependencies [b1a394c]
+  - @slashid/react@1.24.0
+  - demo-form@0.0.58
+
+## 0.0.64
+
+### Patch Changes
+
+- Updated dependencies [6738884]
+  - @slashid/react@1.23.0
+  - demo-form@0.0.57
+
 ## 0.0.63
 
 ### Patch Changes