+++ packages/demo-form/CHANGELOG.md
@@ -1,5 +1,27 @@
 # demo-form
 
+## 0.0.59
+
+### Patch Changes
+
+- Updated dependencies [25960c7]
+  - @slashid/react@1.24.1
+
+## 0.0.58
+
+### Patch Changes
+
+- Updated dependencies [689a2cd]
+- Updated dependencies [b1a394c]
+  - @slashid/react@1.24.0
+
+## 0.0.57
+
+### Patch Changes
+
+- Updated dependencies [6738884]
+  - @slashid/react@1.23.0
+
 ## 0.0.56
 
 ### Patch Changes